Update anatLmer to calculate DFs internally
This was a somewhat interesting choice that may be worth changing. The first implementation used mincLmerOptimizeAndExtract, the workhorse of mincLmer, this implementation fits with good ole fashioned lmer. The reason for the change is that mincLmerEstimateDF required fitting 50 models with lmer in order to extract DF info, so 300 fast models + 50 slow models, is probably within range of 300 slow models time wise, and the simplicity of doing both at once is nice.