Change *EstimateDF to support exotic formulae
In order to do this *Lmer's need to emit the
data argument supplied to the original model.
Then the models are refit with lmerTest using
that original data argument and the formula used
stripped of its environment. Then I was
misinterpretting errors thrown by lmerTest::summary
which actually had to do with non-positive-definite
hessians. So I wrapped the summary calculation in
a tryCatch that returns NA if it fails, signalling
a warning. NA's are dealt with with good ole na.rm