Update to roxygen2 6.0.1 they said...
It'll be fun they said...
But in positive news it means markdown can be used in roxygen blocks which is a very nice bonus. I'm not going to enable it globally, although maybe I should experiment with that. For now chunks can be markdownified with the tag @md somewhere in the block